refactor(utils): rename validators to utils
