+++ src/main/java/vendingmachine/utils/PurchaseValidator.java
@@ -1,4 +1,4 @@
-package vendingmachine.validator;
+package vendingmachine.utils;
 
 import vendingmachine.domain.Merchandise;
